+++ src/primaite/game/agent/observations/software_observation.py
@@ -1,7 +1,7 @@
 # © Crown-owned copyright 2024, Defence Science and Technology Laboratory UK
 from __future__ import annotations
 
-from typing import Dict, Optional
+from typing import Dict, List, Optional
 
 from gymnasium import spaces
 from gymnasium.core import ObsType
@@ -95,7 +95,7 @@ class ApplicationObservation(AbstractObservation, identifier="APPLICATION"):
         If True, applications must be scanned to update the health state. If False, true state is always shown.
         """
 
-    def __init__(self, where: WhereType, applications_requires_scan: bool) -> None:
+    def __init__(self, where: WhereType, applications_requires_scan: bool, thresholds: Optional[Dict] = {}) -> None:
         """
         Initialise an application observation instance.
 
@@ -108,23 +108,49 @@ class ApplicationObservation(AbstractObservation, identifier="APPLICATION"):
         self.applications_requires_scan = applications_requires_scan
         self.default_observation = {"operating_status": 0, "health_status": 0, "num_executions": 0}
 
-        # TODO: allow these to be configured in yaml
-        self.high_threshold = 10
-        self.med_threshold = 5
-        self.low_threshold = 0
+        if thresholds.get("app_executions") is None:
+            self.low_app_execution_threshold = 0
+            self.med_app_execution_threshold = 5
+            self.high_app_execution_threshold = 10
+        else:
+            self._set_application_execution_thresholds(
+                thresholds=[
+                    thresholds.get("app_executions")["low"],
+                    thresholds.get("app_executions")["medium"],
+                    thresholds.get("app_executions")["high"],
+                ]
+            )
+
+    def _set_application_execution_thresholds(self, thresholds: List[int]):
+        """
+        Method that validates and then sets the application execution threshold.
+
+        :param: thresholds: The application execution threshold to validate and set.
+        """
+        if self._validate_thresholds(
+            thresholds=[
+                thresholds[0],
+                thresholds[1],
+                thresholds[2],
+            ],
+            threshold_identifier="app_executions",
+        ):
+            self.low_app_execution_threshold = thresholds[0]
+            self.med_app_execution_threshold = thresholds[1]
+            self.high_app_execution_threshold = thresholds[2]
 
     def _categorise_num_executions(self, num_executions: int) -> int:
         """
-        Represent number of file accesses as a categorical variable.
+        Represent number of application executions as a categorical variable.
 
-        :param num_access: Number of file accesses.
+        :param num_access: Number of application executions.
         :return: Bin number corresponding to the number of accesses.
         """
-        if num_executions > self.high_threshold:
+        if num_executions > self.high_app_execution_threshold:
             return 3
-        elif num_executions > self.med_threshold:
+        elif num_executions > self.med_app_execution_threshold:
             return 2
-        elif num_executions > self.low_threshold:
+        elif num_executions > self.low_app_execution_threshold:
             return 1
         return 0
 
@@ -180,4 +206,5 @@ class ApplicationObservation(AbstractObservation, identifier="APPLICATION"):
         return cls(
             where=parent_where + ["applications", config.application_name],
             applications_requires_scan=config.applications_requires_scan,
+            thresholds=config.thresholds,
         )